--- src/os/version_other.go
@@ -1,11 +1,0 @@
-// Copyright 2024 The Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-//go:build !android
-
-package os
-
-func androidVersion() int {
-	return 0
-}
