+++ src/runtime/sys_libc.go
@@ -14,6 +14,7 @@ import "unsafe"
 // Preserves the calling point as the location where a profiler traceback will begin.
 //
 //go:nosplit
+//go:linkname libcCall runtime.libcCall
 func libcCall(fn, arg unsafe.Pointer) int32 {
 	// Leave caller's PC/SP/G around for traceback.
 	gp := getg()
